DomainToText_AMLProject/requirements.txt
torch==1.4.0
torchvision==0.5.0
future==0.18.2
Pillow==6.2.2
numpy==1.17.0
transformers==4.11.0
tqdm==4.48.2
